split navbar.js into smaller modules
