+++ src/pages/Cart/Delivery.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import Navbar from "../../components/landing-page/Navbar";
+import Navbar from "../../components/Navbar/Navbar";
 import Footer from "../../components/landing-page/Footer";
 import DeliveryForm from '../../components/Delivery/DeliveryForm';
 import "../../styles/CartStyles/Delivery.css";
