+++ src/pages/Cart/EmptyCart.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { Link } from "react-router-dom";
-import Navbar from "../../components/landing-page/Navbar";
+import Navbar from "../../components/Navbar/Navbar";
 import Footer from "../../components/landing-page/Footer";
 import { useProductContext } from "../../context/product_context";
 import Product from "../../components/Product/Product";
