+++ src/pages/StaticPages/Aboutus.js
@@ -1,6 +1,6 @@
 import React, { useState } from 'react'
 import { Link } from "react-router-dom";
-import Navbar from "../../components/landing-page/Navbar";
+import Navbar from "../../components/Navbar/Navbar";
 import Footer from "../../components/landing-page/Footer";
 import "../../styles/StaticStyles/Aboutus.css";
 import aboutimg from "../../assets/StaticPages/About.png";
