+++ src/pages/StaticPages/termConditions.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import Navbar from "../../components/landing-page/Navbar";
+import Navbar from "../../components/Navbar/Navbar";
 import Footer from "../../components/landing-page/Footer";
 import Conditions from '../../components/StaticPages/Condition/Conditions'
 import { conditionObjOne } from '../../data/StaticPages/ConditionsData'
