+++ src/styles/Hero/hero.scss
@@ -1,4 +1,4 @@
-@import '../../styles/landing/variables';
+@import '../variables';
 
 .hero {
   position: relative;
