+++ src/styles/Home/home.scss
@@ -1,4 +1,4 @@
-@import "../landing/variables";
+@import "../variables";
 
 .home__hero {
   height: 88vh;
