+++ src/styles/MarketPlace/marketPlace.scss
@@ -6,7 +6,7 @@
 
 .marketPlace__products{
   display: flex;
-  justify-content: center;
+  justify-content: flex-start;
   align-items: center;
   gap: 2.5rem;
   flex-wrap: wrap;
