+++ src/styles/navbar/mobile_search.scss
@@ -1,0 +1,34 @@
+.mobile__search {
+  position: absolute;
+  top: 10vh;
+  left: 0;
+  height: 20vh;
+  width: 100%;
+  padding: 0 2rem;
+  background: #fff;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  //z-index: -50;
+  opacity: 0;
+  transition: opacity 800ms ease;
+  display: none;
+
+  @media (max-width: 480px) {
+    display: block;
+  }
+
+  input {
+    width: 90%;
+    height: 100%;
+    padding: 1rem 1rem 1rem 1.5rem;
+  }
+
+  img {
+    margin-left: 0.5rem;
+  }
+}
+
+.show__mobileSearch {
+  opacity: 1;
+}
