+++ src/styles/product/addToCart.scss
@@ -1,4 +1,4 @@
-@import '../landing/variables';
+@import '../variables';
 
 .addToCart__button div{
   width: 100%;
